new ppyolo class
